% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-brml2df.R
\name{brml2df}
\alias{brml2df}
\title{Import data and metadata from Bruker BRML files}
\usage{
brml2df(brml.file)
}
\arguments{
\item{brml.file}{complete path with filename to BRML file}
}
\value{
dataframe
}
\description{
The BRML archive is readable with standard zip/unzip tools.
I have found that the BRML archive contains the following files/folders:
+ Experiment0/
+ Experiment0/ActiveContentContainer.xml
+ Experiment0/DataContainer.xml
+ Experiment0/EvaluationContainer.xml
+ Experiment0/InstructionContainer.xml
+ Experiment0/MeasurementContainer.xml
+ Experiment0/TemplateContainer.xml
+ experimentCollection.xml
and depending on the number of measurements, the folder Experiment0 will
contain the same number of RawData#.xml files, numbered starting from zero
+ Experiment0/RawData0.xml
+ Experiment0/RawData1.xml
+ ...
}
