% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-brukeruxd2df.R
\name{brukeruxd2df}
\alias{brukeruxd2df}
\title{Import data and metadata from UXD files}
\usage{
brukeruxd2df(uxdfile)
}
\arguments{
\item{uxdfile}{path to UXD file}
}
\value{
dataframe with the following columns
  $ sampleid          : chr
  $ substrateid       : chr
  $ range             : chr
  $ counts (or cps)   : num
  $ steptime          : num
  $ stepsize          : chr
  $ stepmode          : chr
  $ start2th          : chr
  $ rotationspeed     : chr
  $ voltage           : chr
  $ current           : chr
  $ angle             : num
  $ cps (or counts)   : num
  $ goniometer.radius : num
  $ fixed.divslit     : num
  $ fixed.antislit    : num
  $ datemeasured      : num
  $ anode             : num
}
\description{
Read UXD files from Bruker D8 with one or multiple ranges (converted from
Bruker's native  format using their FileExchange tool).
This function was tested for UXD v3.
It automatically calculates CPS (counts per second) if COUNTS are present,
and vice versa (note that this depends on specific strings in the UXD format).
Old UXD-files (probably from the Siemens diffractometers) contain the following line:
"; (Data for Range number 1)".
Converted UXD files (from the Bruker D8) drop the parentheses and "number":
"; Data for range 1"
This function was tested for the latter case.
}
