% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-mraw2df.R
\name{mraw2df}
\alias{mraw2df}
\title{Import data and metadata from XRD Commander exported TXT files}
\usage{
mraw2df(txtfile)
}
\arguments{
\item{txtfile}{path to TXT file}
}
\value{
dataframe with the following columns:
  $ sampleid        : chr
  $ thth            : num
  $ counts (or cps) : num
  $ steptime        : num
  $ stepsize        : num
  $ theta           : num
  $ khi             : num
  $ phi             : num
  $ x               : num
  $ y               : num
  $ z               : num
  $ divergence      : num
  $ antiscatter     : num
  $ cps (or counts) : num
}
\description{
Reads TXT files with one or multiple ranges (XRD commander "save as txt")
}
