% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-muxd2df.R
\name{muxd2df}
\alias{muxd2df}
\title{Import data and metadata from UXD files converted using the XCH tool}
\usage{
muxd2df(uxdfile)
}
\arguments{
\item{uxdfile}{path to UXD file}
}
\value{
dataframe with the following columns:
  $ sampleid        : chr
  $ substrateid     : chr
  $ thth            : num
  $ counts (or cps) : num
  $ steptime        : num
  $ stepsize        : num
  $ theta           : num
  $ khi             : num
  $ phi             : num
  $ x               : num
  $ y               : num
  $ z               : num
  $ divergence      : num
  $ antiscatter     : num
  $ cps (or counts) : num
}
\description{
Read UXD files from Bruker diffractometer with one or multiple ranges
(converted from Bruker's native format using XCH v1.0).
It automatically calculates CPS (counts per second) if COUNTS are present,
and vice versa (note that this depends on specific strings in the UXD format).
Old UXD-files (probably from the Siemens diffractometers) contain the following line:
"; (Data for Range number 1)"
Converted UXD files (from the Bruker D8) drop the parentheses and "number":
"; Data for range 1"
This function is for the **former** case.
}
