% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-muxd2ls.R
\name{muxd2ls}
\alias{muxd2ls}
\title{Convert UXD file to R list object}
\usage{
muxd2ls(uxdfile)
}
\arguments{
\item{uxdfile}{path to UXD file}
}
\value{
list of matrices, as many as there were ranges in the UXD file
}
\description{
This function reads an UXD file which contains several ranges (created in
a programmed run, for example).
}
