% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-muxd2mtx.R
\name{muxd2mtx}
\alias{muxd2mtx}
\title{Convert UXD file to R matrix object}
\usage{
muxd2mtx(uxdfile, range)
}
\arguments{
\item{uxdfile}{path to UXD file (string)}

\item{range}{number of ranges in UXD file (integer)}
}
\value{
matrix with two columns
}
\description{
This function reads an UXD file which contains several ranges (created in
a programmed run, for example).
}
