% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-uxd2mtx.R
\name{uxd2mtx}
\alias{uxd2mtx}
\title{Convert UXD file containing a single range to R matrix object}
\usage{
uxd2mtx(uxdfile)
}
\arguments{
\item{uxdfile}{path to UXD file}
}
\value{
matrix with two columns
}
\description{
Assumes the UXD file contains two columns.
}
