% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-xrdml2df.R
\name{xrdml2df}
\alias{xrdml2df}
\title{Read data and metadata from XRDML files}
\usage{
xrdml2df(xrdml_source_path)
}
\arguments{
\item{xrdml_source_path}{path to XRDML file}
}
\value{
list with three dataframes:
  xrd.data:
    $ angle
    $ counts
    $ att.factors
    $ secperstep
    $ cps
    $ cpspermA
    $ sampleid
    $ substrateid
    $ comment
    $ usedWavelength
    $ IB.radius
    $ IB.xraytube
    $ IB.xray.anode
    $ IB.xray.tension
    $ IB.xray.current
    $ IB.focus.type
    $ IB.focus.length
    $ IB.focus.width
    $ IB.takeoff.angle
    $ IB.xray.mirror
    $ IB.soller.slit
    $ IB.beam.attenuator
    $ sample.offset
    $ DB.radius
    $ DB.parallel.plate.collimator
    $ DB.detector
    $ scan.mode
    $ scan.axis
    $ start.time
    $ end.time
    $ application.software
    $ control.software
    $ intensities.unit
    $ common.counting.time
  axis.data
    $ sampleid
    $ axis
    $ unit
    $ startPosition
    $ endPosition
    $ commonPosition
  xrdml.params
    $ sampleid
    $ substrateid
    $ comment
    $ usedWavelength
    $ IB.radius
    $ IB.xraytube
    $ IB.xray.anode
    $ IB.xray.tension
    $ IB.xray.current
    $ IB.focus.type
    $ IB.focus.length
    $ IB.focus.width
    $ IB.takeoff.angle
    $ IB.xray.mirror
    $ IB.soller.slit
    $ IB.beam.attenuator
    $ sample.offset
    $ DB.radius
    $ DB.parallel.plate.collimator
    $ DB.detector
    $ scan.mode
    $ scan.axis
    $ start.time
    $ end.time
    $ application.software
    $ control.software
    $ intensities.unit
    $ common.counting.time
}
\description{
Read data and metadata from XRDML files
}
